% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_analysis_expression.R
\name{qPCR_analysis_expression}
\alias{qPCR_analysis_expression}
\title{Real-time qPCR relative quantity calculation.}
\usage{
qPCR_analysis_expression(unkdata, refgene, efficiencies = NULL, control = NULL)
}
\arguments{
\item{unkdata}{A \code{data.frame} containing the technical replicate means of the unknown reactions.
Result from \code{qPCR_analysis_unk_rxns} function.}

\item{refgene}{Character vector with the name(s) of the reference gene(s).}

\item{efficiencies}{A \code{data.frame} containing the efficiencies of each primer pair.
The \verb{$efficiencies} \code{data.table} resulting from the \code{qPCR_analysis_stdcurve} function.
Defaults to \code{NULL}, which assumes all primer pairs' efficiency being equal to 100\%.}

\item{control}{Character. The name of the control sample, against which all other samples will be compared.
Defaults to the 1st sample alphabetically.}
}
\value{
A \code{data.table} with the calculated expression of each sample, biological replicate and target gene.
}
\description{
Calculates the relative quantity of target genes between one control sample and one or more test samples.
Based on one or multiple reference genes.
}
\details{
Applies the Pfaffl calculation method, accounting for different primer efficiencies
(\href{https://doi.org/10.1093/nar/29.9.e45}{Pfaffl, 2001}).
The common-base calculation method described by \href{https://doi.org/10.1186/s12859-017-1949-5}{Ganger et al. 2017} is followed,
which gives identical results to the Pfaffl method.
If efficiencies are not entered (\code{NULL}), it defaults to the delta-delta-Ct method
(\href{https://doi.org/10.1006/meth.2001.1262}{Livak & Schmittgen, 2001}).
}
