% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_analysis_ntc.R
\name{qPCR_analysis_ntc}
\alias{qPCR_analysis_ntc}
\title{Real-time qPCR non-template control processing.}
\usage{
qPCR_analysis_ntc(d)
}
\arguments{
\item{d}{\code{data.frame}, Cq value data exported from Bio-Rad CFX Connect real-time PCR machine.}
}
\value{
A \code{data.table} with primer targets and Cq values, each row being a non-template control reaction.
}
\description{
Parses primary qPCR data (Cq values) and summarizes results for non-template controls.
Compatible with Cq data exported from Bio-Rad CFX Connect real-time PCR machine.
}
