% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_analysis_stdcurve.R
\name{qPCR_analysis_std_curve}
\alias{qPCR_analysis_std_curve}
\title{Real-time qPCR standard curve and primer efficiency calculation.}
\usage{
qPCR_analysis_std_curve(d, graph = FALSE)
}
\arguments{
\item{d}{\code{data.frame}, Cq value data exported from Bio-Rad CFX Connect real-time PCR machine.}

\item{graph}{Logical. If \code{TRUE}, also draws and returns the standard curves on a scatter plot with trend lines. Defaults to \code{FALSE}.}
}
\value{
A list with 3 elements:
\itemize{
\item \verb{$data}, a \code{data.table} with the Cq values against log(Starting quantity) for each target gene.
\item \verb{$efficiencies}, a \code{data.table} with the calculated efficiencies for each target gene.
\item \verb{$graph}, the scatter plot with the drawn standard curves if requested.
Points are technical means, error bars are technical standard deviations.
}
}
\description{
Parses primary qPCR data (Cq values) and calculates primer efficiencies based on standard curve serial dilutions.
Compatible with Cq data exported from Bio-Rad CFX Connect real-time PCR machine.
}
\details{
Standard curve slopes and resulting efficiencies are calculated by fitting the Cq values (y) against the log(starting quantity) (x)
on a linear model (function \code{lm(y~x)}).
}
