% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_analysis_unk_rxns.R
\name{qPCR_analysis_unk_rxns}
\alias{qPCR_analysis_unk_rxns}
\title{Real-time qPCR unknown sample reaction processing.}
\usage{
qPCR_analysis_unk_rxns(d)
}
\arguments{
\item{d}{\code{data.frame}, Cq value data exported from Bio-Rad CFX Connect real-time PCR machine.}
}
\value{
A \code{data.table} with Cq values for each sample, biological replicate and gene. Contains the following fields:
\itemize{
\item \code{Sample}: Name of sample template.
\item \code{Biol.rep}: Biological replicate number.
\item \code{Target}: Name of target gene.
\item \code{Cq.tech.mean}: Arithmetic mean between technical replicates.
\item \code{Cq.tech.sd}: Standard deviation between technical replicates.
}
}
\description{
Parses primary qPCR data (Cq values) and calculates technical means for each sample, biological replicate and gene.
Compatible with Cq data exported from Bio-Rad CFX Connect real-time PCR machine.
}
