% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_analysis_wrap.R
\name{qPCR_analysis_wrap}
\alias{qPCR_analysis_wrap}
\title{Real-time qPCR result processing.}
\usage{
qPCR_analysis_wrap(
  d,
  refgene,
  control = NULL,
  std.curve = TRUE,
  std.curve.plot = TRUE
)
}
\arguments{
\item{d}{\code{data.frame}, Cq value data exported from Bio-Rad CFX Connect real-time PCR machine.}

\item{refgene}{Character vector with the name(s) of the reference gene(s).}

\item{control}{Character. The name of the control sample, against which all other samples will be compared.
Defaults to the 1st sample alphabetically.}

\item{std.curve}{Logical. Determines whether or not to calculate efficiencies from standard curves or not.
If \code{FALSE}, standard curve reactions will be ignored and efficiencies of 100\% will be assumed. Defaults to \code{TRUE}.}

\item{std.curve.plot}{Logical. Determines whether or not to plot standard curve data in a scatter plot with trend lines.
Defaults to \code{TRUE}.}
}
\value{
A list of 4 elements:
\itemize{
\item \verb{$NTC}: Results from \code{qPCR_analysis_ntc} function.
\item \verb{$std.curve}: Results from \code{qPCR_analysis_std_curve} function.
\item \verb{$unk.rxn}: Results from \code{qPCR_analysis_unk_rxns} function.
\item \verb{$expression}: Results from \code{qPCR_analysis_expression} function.
}
}
\description{
Performs the whole processing of real-time qPCR data (Cq values),
including non-template control summary, standard curve analysis and expression analysis.
Compatible with Cq data exported from Bio-Rad CFX Connect real-time PCR machine.
}
